Splitted dev-tests utils into two directories.
