+++ packages/ckeditor5-dev-tests/lib/index.js
@@ -6,5 +6,5 @@
 module.exports = {
 	runAutomatedTests: require( './tasks/runautomatedtests' ),
 	runManualTests: require( './tasks/runmanualtests' ),
-	parseArguments: require( './utils/parsearguments' ),
+	parseArguments: require( './utils/automated-tests/parsearguments' ),
 };
