+++ packages/ckeditor5-dev-tests/lib/tasks/runautomatedtests.js
@@ -7,7 +7,7 @@
 
 'use strict';
 
-const getKarmaConfig = require( '../utils/getkarmaconfig' );
+const getKarmaConfig = require( '../utils/automated-tests/getkarmaconfig' );
 const KarmaServer = require( 'karma' ).Server;
 const path = require( 'path' );
 const gutil = require( 'gulp-util' );
