+++ packages/ckeditor5-dev-tests/lib/utils/automated-tests/getwebpackconfigforautomatedtests.js
