+++ packages/ckeditor5-dev-tests/lib/utils/manual-tests/createmanualtestserver.js
@@ -10,7 +10,7 @@ const path = require( 'path' );
 const fs = require( 'fs' );
 const combine = require( 'dom-combiner' );
 const { logger } = require( '@ckeditor/ckeditor5-dev-utils' );
-const globSync = require( './glob' );
+const globSync = require( '../glob' );
 
 // TODO this bit repeats in 3 files.
 const escapedPathSep = path.sep == '/' ? '/' : '\\\\';
