+++ packages/ckeditor5-dev-tests/lib/utils/manual-tests/getwebpackconfigformanualtests.js
