+++ packages/ckeditor5-dev-tests/tests/tasks/runmanualtests.js
@@ -28,9 +28,9 @@ describe( 'runManualTests', () => {
 		const htmlFileCompilerSpy = sandbox.spy( () => Promise.resolve() );
 		const scriptCompilerSpy = sandbox.spy( () => Promise.resolve() );
 
-		mockery.registerMock( '../utils/createmanualtestserver', serverSpy );
-		mockery.registerMock( '../utils/compilemanualtesthtmlfiles', htmlFileCompilerSpy );
-		mockery.registerMock( '../utils/compilemanualtestscripts', scriptCompilerSpy );
+		mockery.registerMock( '../utils/manual-tests/createmanualtestserver', serverSpy );
+		mockery.registerMock( '../utils/manual-tests/compilemanualtesthtmlfiles', htmlFileCompilerSpy );
+		mockery.registerMock( '../utils/manual-tests/compilemanualtestscripts', scriptCompilerSpy );
 
 		sandbox.stub( path, 'join', ( ...chunks ) => chunks.join( '/' ) );
 		sandbox.stub( process, 'cwd', () => 'workspace' );
